m10/semana_7/Classificação de Imagens (CIFAR-10)/requirements.txt
Flask
Pillow
torch
torchvision
d2l